+++ test/core-e2e-legacy/src/fixtures/config-file-with-plugins/mocks-server-with-plugins.config.js
@@ -1,0 +1,15 @@
+const path = require("path");
+const TraceRoutesPlugin = require("./TraceRoutesPlugin");
+
+module.exports = {
+  log: "silly",
+  files: {
+    path: path.resolve(__dirname, "..", "temp"),
+  },
+  plugins: {
+    register: [TraceRoutesPlugin],
+  },
+  mocks: {
+    selected: "user-2",
+  },
+};