+++ test/core-e2e/src/fixtures/custom-routes-handler/routes/users.js
@@ -18,6 +18,7 @@ module.exports = [
     variants: [
       {
         id: "success",
+        handler: "json",
         response: {
           status: 200,
           body: USERS,