+++ test/core-e2e/src/fixtures/validation-not-array/routes/routes-valid.js
@@ -6,6 +6,7 @@ module.exports = [
     variants: [
       {
         id: "success",
+        handler: "json",
         response: {
           status: 200,
           body: {},